Going out to test GPS

+++ Sensor_Fusion/Core/Inc/SF_Nav/SFNav.h
@@ -26,7 +26,7 @@ public:
 	SF_Nav();
 	virtual ~SF_Nav();
 
-	void init(UART_HandleTypeDef* uh, I2C_HandleTypeDef* ih,float refresh_time);
+	void init(UART_HandleTypeDef* uh, I2C_HandleTypeDef* ih,int refresh_time);
 
 	void update();
 	state_var get_state();
